Update main.go
Update parseTrailCamData method to handle extracting camera name when there is a prefix of a single letter and a space. This is due to the OCR recognizing the symbol representing the moon phase as a letter. Also, changed the Temperature field of the TrailCamData struct to be an integer since the OCR always returns whole numbers. Updated the log statement to show the temperature unit as a string instead of a memory address.