Refactoring the frame processing
Optimizing the frame processing by creating a new label image that will have the cropped bounding box overlayed onto it.

The label images are then concatenated together into a single.
That single image is passed to performOCR and we get back a text string that is easy to parse into key value pairs.

These changes nearly double the overall processing speed.

Also fixing the readFrame function so that we can advance the frame number by percentage in cases when reading frame 0 fails.